[TASK] Add JSON data wrapper
This encapsulates JSON parsing and exposes the parsed data as array-ish
object.

Raw data from the database is passed in as constructor argument by the
Extbase DataMapper thanks to the core TypeInterface. The same interface
could be used to format updated data for persistence but this is not
used here.