[BUGFIX] Fix form setup loading in logger finisher functional tests
The form setup is loaded by YamlFileLoader which resolves relative
import paths with its getStreamlinedFileName() method. This method uses
GeneralUtility::isAllowedAbsPath() to determine if the resolved file
path is within the web root. This is not the case for functional
test instances where the "typo3" directory is only a symlink to the
actual directory. Due to this all relative imports within the
FormSetup.yaml are skipped.

Thankfully GeneralUtility::isAllowedAbsPath() also allowsfiles in the
barely known "lockRootPath". By setting this to the original root, the
relative imports are working.

See https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/GlobalValues/Typo3ConfVars/BE.html#globals-typo3-conf-vars-be-lockrootpath