[TASK] Use subrequest for functional LoggerFinisher tests
Use a regular subrequest to fetch and submit forms with data.

This allows us to drop all kinds of mocking and fake environment setup.
Also we do not need to know about the internals of the Form framework
anymore.

The hardest part was testing file uploads which evidently do not work
with stream-based uploaded files. When using the StreamFactoryInterface
to create one and "upload" such a file, Extbase/Form will not accept
this file since there is no "tmp_name" internally. This only works with
phyisically created temporary files just like we did before.